backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.1

openai==1.40.6

azure-ai-textanalytics==5.3.0
azure-core==1.30.2
azure-cognitiveservices-speech==1.40.0
azure-identity==1.15.0

numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
plotly==5.22.0